+++ src/main/java/org/scoula/domain/chatbot/dto/ChatMessageDto.java
@@ -15,9 +15,9 @@ import java.time.LocalDateTime;
 @Builder 
 @ApiModel(discriminator = "메시지 저장/조회 dto")
 public class ChatMessageDto {
-    private Long id; // 메세지 ID
-    private Long userId; //유저 ID
-    private Long sessionId; // 세션 ID
+    private Integer id; // 메세지 ID
+    private Integer userId; //유저 ID
+    private Integer sessionId; // 세션 ID
     private String role;  // "user" / "assistant"
     private String content;  // 메세지 내용
     private String intentType;  // optional
